modify MakeKey format

+++ include/AugeType.h
@@ -16,13 +16,13 @@ typedef unsigned long		g_ulong;
 #if defined(_MSC_VER) || defined(__BORLANDC__)
 typedef signed __int64      g_int64;
 #else
-typedef signed long long    g_int64;
+typedef signed long int    g_int64;
 #endif
 
 #if defined(_MSC_VER) || defined(__BORLANDC__)
 typedef unsigned __int64    g_uint64;
 #else
-typedef unsigned long long  g_uint64;
+typedef unsigned long int  g_uint64;
 #endif
 
 #ifndef PI
